pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jera-cli"
version = "1.0.0"
description = "CLI simplificada para gerenciar recursos da Jera na AWS"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "click==8.1.7",
    "kubernetes==29.0.0",
    "rich==13.7.0",
    "inquirer==3.1.3",
    "PyYAML==6.0.1",
]

[tool.setuptools]
py-modules = ["jera_cli"] 